Fix for nasty I2C eeprom data corruption bug
Writing to I2C eeprom with 6 bytes of data was triggering an
intermittent timing bug where  two byte pairs would be swapped with
another two of the six bytes. Happened every 21 writes of 6 bytes (with
some rare fluctuations). Writing the data one byte at a time, or in
pairs resolved the issue.