Major re-write for SMT v0.6 Rev A board
Changes to much of the code to support the new TCS34725FN I2C light
sensor device, and the new 5 LEDs for user bar graph goal feedback. The
new sensor provides 16bit per channel of data  and is currently set to
cover 1 lux to 100+ klux. Code includes some bit tricks to reduce that
to 10bits per channel at the cost of reduced resolution as the lux
increases to large values (similar to the behaviour of the previous LDR
sensors). The TCS34725FN device is using low power management between
readings, but it may be better to fully power the device down.
TCS34725FN also supports changing integration time and sensor gain,
these may be used in future build to extend the light sensitivity above
and below the current range.