Refactor out gosubs and add delay compensation
After monitoring timing and power rails via oscilloscope it’s clear
that gosub/returns are rather time expensive, this can be a hit on the
power usage if they are happening during higher power modes. This patch
removes many of the gosubs, saves a little memory, and adds in a
calculation to allow the customisable delay value to be consistent
between firmware updates even if the code path is longer or shorter
than previously.